requirements.txt
langchain>=0.3.25
langchain-community>=0.3.24
langchain-core>=0.3.61
langchain-text-splitters>=0.3.8
langgraph>=0.4.6
langgraph-checkpoint>=2.0.26
langgraph-prebuilt>=0.2.1
langsmith>=0.3.42
ollama>=0.1.7
python-dotenv>=1.0.0
